requirements.txt
requests
pandas
streamlit
plotly
sqlalchemy
python-dateutil
